Store port addresses and shape counts as integers (#102)
* Add converter package: Converter and ElementsConverter base classes

Convert domain-specific objects (e.g. pypowsybl networks) into
energnn.graph.Graph instances: string addresses are mapped to
consecutive integers in deterministic sorted order, features are cast
to bounded floats (categorical values hashed to a stable float, NaN
filled, inf clipped), and graphs are assembled through
HyperEdgeSet.from_dict/Graph.from_dict on a numpy backend, with an
optional target backend conversion.


* Use converters to build graphs in the LinearSystem example

Replace the manual HyperEdgeSet/Graph construction in generate_problem
with LinearSystemContextConverter and LinearSystemOracleConverter, and
derive the context/decision structures from the converters instead of
duplicating them. Oracles carry only features: they now have an empty
address registry (n_addresses=0) instead of an artificial one.


* Add tests for the converter package and LinearSystem converters

Cover the ElementsConverter port/feature split and guards, the core
conversion invariants (consistent integer address mapping, determinism,
input non-mutation, backend handling, pad/collate compatibility), the
_str_to_int and _any_to_float helpers, and the LinearSystem example
converters.


* Expose converter package at top level and add its API reference docs


* Vectorize address mapping and categorical hashing in converters

Address review feedback on the converter pipeline performance:
- _str_to_int now collects distinct addresses with a hash table
  (pd.unique), sorts only the uniques, and translates tables through
  pd.Index.get_indexer instead of a per-element Python lambda.
- _any_to_float hashes each distinct category once and broadcasts the
  result with Series.map instead of seeding a PRNG per row. NaNs in
  categorical columns now fall through to fillna(0) like numeric NaNs,
  as documented.

The integer mapping and hashed values are unchanged.


* Store port addresses and shape counts as integers

Ports and shape counts were stored as float32 everywhere: to_numpy forced
float32 at creation, and Backend.from_numpy re-cast them to the backend's
floating dtype on every to_backend call, compensated downstream by
astype(int) at gather/scatter sites. This silently corrupts addresses
beyond 2^24, and beyond 256 with a bfloat16 backend.

Ports and shape counts are now int32 end to end:

- HyperEdgeSet.from_dict casts ports to int32 and features to float32;
  integer-valued float ports are still accepted, fractional or NaN ports
  are rejected.
- Backend.from_numpy accepts a PRESERVE_DTYPE sentinel so that to_backend
  never routes integer data (ports, shape counts, feature-name indices)
  through the backend's floating dtype.
- to_numpy preserves dtypes instead of forcing float32.
- GraphShape counts are built as int32; GraphShape.max defaults missing
  classes to 0 instead of -inf to avoid float promotion.
- Compensatory astype(int) casts in gather/scatter call sites are removed.

Benchmarked on a jitted grad step through LocalSumMessagePassingFunction:
no measurable difference in time or peak memory vs float32 ports — the
change is about correctness, not speed.


---------

+++ src/energnn/graph/graph.py
@@ -291,15 +291,15 @@ class Graph(dict):
             for edge_key, edge in graph.hyper_edge_sets.items():
                 edge_h[edge_key] = []
                 for _, address_array in edge.port_dict.items():
-                    edge_h[edge_key].append(h_new_[address_array.astype(int)])
+                    edge_h[edge_key].append(h_new_[address_array])
                 edge_h[edge_key] = xp.stack(edge_h[edge_key], axis=0)
                 edge_h[edge_key] = xp.max(edge_h[edge_key], axis=0)
                 for _, address_array in edge.port_dict.items():
                     new_val = xp.max(
-                        xp.stack([edge_h[edge_key], h_new_[address_array.astype(int)]], axis=0),
+                        xp.stack([edge_h[edge_key], h_new_[address_array]], axis=0),
                         axis=0,
                     )
-                    h_new_ = backend.scatter_max(h_new_, address_array.astype(int), new_val)
+                    h_new_ = backend.scatter_max(h_new_, address_array, new_val)
             return h_new_
 
         if not self.is_single:
